refactor: split sync_event_processor.dart
