+++ lib/features/sync/matrix/sync_event_processor.dart
@@ -46,6 +46,13 @@ export 'package:lotti/features/sync/matrix/smart_journal_entity_loader.dart';
 export 'package:lotti/features/sync/matrix/sync_journal_entity_loader.dart';
 export 'package:lotti/features/sync/matrix/vector_clock_validator.dart';
 
+// Per-domain method bodies live in part files so they share the orchestrator's
+// private state (dedup cache, sequence log service, agent repository) without
+// dependency-injection plumbing for every collaborator.
+part 'sync_event_processor_journal_handlers.dart';
+part 'sync_event_processor_agent_handlers.dart';
+part 'sync_event_processor_outbox_bundle.dart';
+
 /// Sync message bodies below this base64 length are decoded inline; anything
 /// longer hands off to a worker isolate via `compute`. The break-even point
 /// covers the `compute` spin-up cost on typical desktop hardware — small
@@ -305,148 +312,6 @@ class SyncEventProcessor {
     return message.contains('stale attachment json');
   }
 
-  /// Upserts every embedded entry link and returns the count that
-  /// actually wrote a row.
-  ///
-  /// Set [rethrowOnError] when this runs inside a transaction whose
-  /// commit must be conditional on every link succeeding — e.g. the
-  /// journal-entity apply path, where the entity row + its embedded
-  /// links need to land atomically. Without rethrow, a single
-  /// `upsertEntryLink` failure would be logged and the loop would
-  /// continue, then the transaction would commit the entity *with*
-  /// only some of its links; later redelivery of the same event
-  /// resolves to the duplicate path and never retries the missing
-  /// upserts.
-  ///
-  /// Default `false` preserves the legacy behaviour for callers that
-  /// process links *outside* a transaction (best-effort delivery), so
-  /// a poison link does not block unrelated work.
-  Future<int> _processEmbeddedEntryLinks({
-    required List<EntryLink>? entryLinks,
-    required JournalDb journalDb,
-    bool rethrowOnError = false,
-  }) async {
-    var processedLinksCount = 0;
-    if (entryLinks == null || entryLinks.isEmpty) {
-      return processedLinksCount;
-    }
-    final affectedIds = <String>{};
-    for (final link in entryLinks) {
-      try {
-        final linkRows = await journalDb.upsertEntryLink(link);
-        if (linkRows > 0) {
-          processedLinksCount++;
-          _trace(
-            'apply entryLink.embedded from=${link.fromId} to=${link.toId} rows=$linkRows',
-            subDomain: 'processor.apply.entryLink.embedded',
-          );
-        }
-        affectedIds.addAll({link.fromId, link.toId});
-      } catch (e, st) {
-        _loggingService.captureException(
-          e,
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'apply.entryLink.embedded',
-          stackTrace: st,
-        );
-        if (rethrowOnError) rethrow;
-      }
-    }
-    if (affectedIds.isNotEmpty) {
-      _updateNotifications.notify(affectedIds, fromSync: true);
-    }
-    return processedLinksCount;
-  }
-
-  Future<SyncApplyDiagnostics?> _maybeSkipSupersededStaleDescriptor({
-    required Event event,
-    required SyncJournalEntity syncMessage,
-    required JournalDb journalDb,
-    required List<EntryLink>? entryLinks,
-  }) async {
-    final incomingVc = syncMessage.vectorClock;
-    if (incomingVc == null) {
-      return null;
-    }
-    JournalEntity? existing;
-    try {
-      existing = await journalDb.journalEntityById(syncMessage.id);
-    } catch (e, st) {
-      _loggingService.captureException(
-        e,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'apply.staleDescriptor.lookup',
-        stackTrace: st,
-      );
-      return null;
-    }
-    final existingVc = existing?.meta.vectorClock;
-    if (existingVc == null) {
-      return null;
-    }
-    VclockStatus status;
-    try {
-      status = VectorClock.compare(existingVc, incomingVc);
-    } catch (e, st) {
-      _loggingService.captureException(
-        e,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'apply.staleDescriptor.compare',
-        stackTrace: st,
-      );
-      return null;
-    }
-    if (status != VclockStatus.a_gt_b && status != VclockStatus.equal) {
-      return null;
-    }
-
-    final processedLinksCount = await _processEmbeddedEntryLinks(
-      entryLinks: entryLinks,
-      journalDb: journalDb,
-    );
-
-    final diag = SyncApplyDiagnostics(
-      eventId: event.eventId,
-      payloadType: 'journalEntity',
-      vectorClock: incomingVc.toJson(),
-      conflictStatus: status.toString(),
-      applied: false,
-      skipReason: JournalUpdateSkipReason.olderOrEqual,
-    );
-    _trace(
-      'apply journalEntity skipped staleDescriptor eventId=${event.eventId} id=${syncMessage.id} status=${diag.conflictStatus} embeddedLinks=$processedLinksCount/${entryLinks?.length ?? 0}',
-      subDomain: 'processor.apply',
-    );
-
-    if (_sequenceLogService != null && syncMessage.originatingHostId != null) {
-      try {
-        final gaps = await _sequenceLogService.recordReceivedEntry(
-          entryId: syncMessage.id,
-          vectorClock: incomingVc,
-          originatingHostId: syncMessage.originatingHostId!,
-          coveredVectorClocks: syncMessage.coveredVectorClocks,
-          jsonPath: syncMessage.jsonPath,
-        );
-        if (gaps.isNotEmpty) {
-          _trace(
-            'apply.gapsDetected count=${gaps.length} for entity=${syncMessage.id}',
-            subDomain: 'processor.gapDetection',
-          );
-        }
-      } catch (e, st) {
-        _loggingService.captureException(
-          e,
-          domain: 'SYNC_SEQUENCE',
-          subDomain: 'recordReceived',
-          stackTrace: st,
-        );
-      }
-    }
-
-    _dedupCache.markProcessed(syncMessage.id, incomingVc);
-    return diag;
-  }
-
   // ---------------------------------------------------------------------------
   // Prepare-phase helpers (I/O only, no DB writes)
   // ---------------------------------------------------------------------------
@@ -580,891 +445,6 @@ class SyncEventProcessor {
     }
   }
 
-  /// Stale-descriptor [FileSystemException]s are caught and carried into the
-  /// apply phase: the supersession check needs the writer transaction, so
-  /// apply decides whether to skip the event or rethrow for retry.
-  Future<PreparedSyncEvent> _prepareJournalEntity({
-    required Event event,
-    required SyncJournalEntity syncMessage,
-  }) async {
-    if (_dedupCache.isDuplicate(syncMessage.id, syncMessage.vectorClock)) {
-      return PreparedSyncEvent._(
-        event: event,
-        syncMessage: syncMessage,
-        isDuplicateJournalEntity: true,
-      );
-    }
-
-    try {
-      final journalEntity = await _journalEntityLoader.load(
-        jsonPath: syncMessage.jsonPath,
-        incomingVectorClock: syncMessage.vectorClock,
-      );
-      return PreparedSyncEvent._(
-        event: event,
-        syncMessage: syncMessage,
-        journalEntity: journalEntity,
-      );
-    } on FileSystemException catch (error, stackTrace) {
-      if (_isStaleDescriptorError(error)) {
-        // Carry the error forward so apply can first check whether the local
-        // version already dominates the incoming one (in which case the
-        // event is skipped) or must be retried later (rethrown from apply).
-        return PreparedSyncEvent._(
-          event: event,
-          syncMessage: syncMessage,
-          deferredStaleDescriptorError: error,
-        );
-      }
-      _loggingService.captureException(
-        error,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SyncEventProcessor.missingAttachment',
-        stackTrace: stackTrace,
-      );
-      rethrow;
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Apply-phase handlers (pure DB, no attachment I/O)
-  // ---------------------------------------------------------------------------
-
-  /// Applies an already-[PreparedSyncEvent] to local stores. Runs entirely in
-  /// the writer transaction — it must not do attachment I/O.
-  Future<SyncApplyDiagnostics?> _applyJournalEntity({
-    required Event event,
-    required SyncJournalEntity syncMessage,
-    required JournalEntity? preloaded,
-    required bool isDuplicate,
-    required FileSystemException? deferredStaleError,
-    required JournalDb journalDb,
-  }) async {
-    if (deferredStaleError != null) {
-      final skipped = await _maybeSkipSupersededStaleDescriptor(
-        event: event,
-        syncMessage: syncMessage,
-        journalDb: journalDb,
-        entryLinks: syncMessage.entryLinks,
-      );
-      if (skipped != null) {
-        return skipped;
-      }
-      // Not superseded — log and rethrow for retry.
-      _loggingService.captureException(
-        deferredStaleError,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SyncEventProcessor.missingAttachment',
-      );
-      throw deferredStaleError;
-    }
-
-    if (isDuplicate) {
-      return _recordDuplicateJournalEntity(
-        event: event,
-        syncMessage: syncMessage,
-      );
-    }
-
-    // preloaded must be non-null when we reach this branch.
-    return _persistJournalEntity(
-      event: event,
-      syncMessage: syncMessage,
-      journalEntity: preloaded!,
-      journalDb: journalDb,
-    );
-  }
-
-  /// Records the duplicate in the sequence log so `resolvePendingHints` still
-  /// runs; the duplicate detection itself lives in the prepare phase.
-  Future<SyncApplyDiagnostics?> _recordDuplicateJournalEntity({
-    required Event event,
-    required SyncJournalEntity syncMessage,
-  }) async {
-    // Even for duplicates, record in the sequence log so that
-    // resolvePendingHints runs. Without this, backfill hints (from
-    // BackfillResponse messages) are never resolved because the entity
-    // already exists locally with the same VC, but the pending
-    // (hostId, counter) → payloadId mapping was never verified.
-    if (_sequenceLogService != null &&
-        syncMessage.vectorClock != null &&
-        syncMessage.originatingHostId != null) {
-      try {
-        await _sequenceLogService.recordReceivedEntry(
-          entryId: syncMessage.id,
-          vectorClock: syncMessage.vectorClock!,
-          originatingHostId: syncMessage.originatingHostId!,
-          coveredVectorClocks: syncMessage.coveredVectorClocks,
-          jsonPath: syncMessage.jsonPath,
-        );
-      } catch (e, st) {
-        _loggingService.captureException(
-          e,
-          domain: 'SYNC_SEQUENCE',
-          subDomain: 'duplicateRecord',
-          stackTrace: st,
-        );
-      }
-    }
-
-    final diag = SyncApplyDiagnostics(
-      eventId: event.eventId,
-      payloadType: 'journalEntity',
-      vectorClock: syncMessage.vectorClock?.toJson(),
-      conflictStatus: VclockStatus.equal.toString(),
-      applied: false,
-      skipReason: JournalUpdateSkipReason.olderOrEqual,
-    );
-    _trace(
-      'apply journalEntity skipped duplicate eventId=${event.eventId} '
-      'id=${syncMessage.id}',
-      subDomain: 'processor.apply',
-    );
-    return diag;
-  }
-
-  /// Persists a pre-resolved journal entity and its embedded links. The
-  /// [journalEntity] argument was already loaded by the prepare phase, so
-  /// this runs entirely in the writer transaction.
-  Future<SyncApplyDiagnostics?> _persistJournalEntity({
-    required Event event,
-    required SyncJournalEntity syncMessage,
-    required JournalEntity journalEntity,
-    required JournalDb journalDb,
-  }) async {
-    final entryLinks = syncMessage.entryLinks;
-
-    // (1) PRE-READ — runs *outside* any journal transaction.
-    //
-    // `applyObserver` only consumes this for diagnostics, and the read
-    // is served by the entity-by-id coalescer which can join an
-    // in-flight wave. Holding the journal writer lock while waiting on
-    // a queued read was one of the worst contention sources in the
-    // super-slow log; the wrap had been added unconditionally by the
-    // queue adapter, so this read was effectively serialized behind
-    // every prior concurrent reader. Now this happens before we open
-    // the inner write transaction.
-    var predictedStatus = VclockStatus.b_gt_a;
-    if (applyObserver != null) {
-      try {
-        final existing = await journalDb.journalEntityById(
-          journalEntity.meta.id,
-        );
-        final vcA = existing?.meta.vectorClock;
-        final vcB0 = journalEntity.meta.vectorClock;
-        if (vcA != null && vcB0 != null) {
-          predictedStatus = VectorClock.compare(vcA, vcB0);
-        }
-      } catch (e, st) {
-        _loggingService.captureException(
-          e,
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'apply.predictVectorClock',
-          stackTrace: st,
-        );
-        predictedStatus = VclockStatus.b_gt_a;
-      }
-    }
-
-    // (2) WRITE — narrow journal transaction wrapping ONLY the writes
-    //     that need to succeed-or-fail together. The queue adapter no
-    //     longer wraps `SyncJournalEntity` from the outside (see
-    //     `_writesJournalDb` in queue_apply_adapter.dart), so this is
-    //     the *only* journal writer-lock acquisition for this apply.
-    //     Cross-DB awaits (sync-sequence log) and the second
-    //     entry-exists read both run *after* this commits, freeing the
-    //     writer lock for unrelated readers.
-    final vcB = journalEntity.meta.vectorClock;
-    final JournalUpdateResult updateResult;
-    final int processedLinksCount;
-    try {
-      final tx = await journalDb.transaction(() async {
-        final result = await journalDb.updateJournalEntity(journalEntity);
-        // Process embedded entry links regardless of journal entity
-        // application status. EntryLinks have their own vector clock
-        // for conflict resolution via upsertEntryLink(). This ensures
-        // links are established even when the entity itself is skipped
-        // (e.g., local version is newer), preventing gray calendar
-        // entries that rely on links for category color lookup.
-        // `rethrowOnError: true` so a failed `upsertEntryLink` aborts
-        // the enclosing transaction. Without that, the entity row
-        // would commit with only a partial link set, and the next
-        // redelivery of this event resolves to the duplicate path —
-        // those missing links would never be retried.
-        final processed = await _processEmbeddedEntryLinks(
-          entryLinks: entryLinks,
-          journalDb: journalDb,
-          rethrowOnError: true,
-        );
-        return (result, processed);
-      });
-      updateResult = tx.$1;
-      processedLinksCount = tx.$2;
-    } catch (error, stackTrace) {
-      _loggingService.captureException(
-        error,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'apply.persistJournalEntity',
-        stackTrace: stackTrace,
-      );
-      rethrow;
-    }
-    final rows = updateResult.rowsWritten ?? 0;
-
-    final diag = SyncApplyDiagnostics(
-      eventId: event.eventId,
-      payloadType: 'journalEntity',
-      vectorClock: vcB?.toJson(),
-      conflictStatus: predictedStatus.toString(),
-      applied: updateResult.applied,
-      skipReason: updateResult.skipReason,
-    );
-    _trace(
-      'apply journalEntity eventId=${event.eventId} id=${journalEntity.meta.id} '
-      'rowsWritten=$rows applied=${updateResult.applied} '
-      'skip=${updateResult.skipReason?.label ?? 'none'} '
-      'status=${diag.conflictStatus} '
-      'embeddedLinks=$processedLinksCount/${entryLinks?.length ?? 0}',
-      subDomain: 'processor.apply',
-    );
-    _dedupCache.markProcessed(
-      journalEntity.meta.id,
-      vcB ?? syncMessage.vectorClock,
-    );
-    _updateNotifications.notify(
-      {...journalEntity.affectedIds, labelUsageNotification},
-      fromSync: true,
-    );
-
-    // (3) POST — sequence-log gap detection. Writes to *sync_db* (a
-    //     separate SQLite database), so it cannot be atomic with the
-    //     journal write in the first place; pulling it out of the
-    //     journal transaction loses no consistency we ever had. Gap
-    //     detection re-runs on next launch, so a crash between (2)
-    //     and (3) is recoverable.
-    if (_sequenceLogService != null &&
-        syncMessage.vectorClock != null &&
-        syncMessage.originatingHostId != null) {
-      final entryExistsInJournal =
-          updateResult.applied ||
-          await journalDb.journalEntityById(journalEntity.meta.id) != null;
-      if (entryExistsInJournal) {
-        try {
-          final gaps = await _sequenceLogService.recordReceivedEntry(
-            entryId: journalEntity.meta.id,
-            vectorClock: syncMessage.vectorClock!,
-            originatingHostId: syncMessage.originatingHostId!,
-            coveredVectorClocks: syncMessage.coveredVectorClocks,
-            jsonPath: syncMessage.jsonPath,
-          );
-          if (gaps.isNotEmpty) {
-            _trace(
-              'apply.gapsDetected count=${gaps.length} '
-              'for entity=${journalEntity.meta.id}',
-              subDomain: 'processor.gapDetection',
-            );
-          }
-        } catch (e, st) {
-          _loggingService.captureException(
-            e,
-            domain: 'SYNC_SEQUENCE',
-            subDomain: 'recordReceived',
-            stackTrace: st,
-          );
-        }
-      }
-    }
-
-    return diag;
-  }
-
-  /// Handles a SyncEntryLink message.
-  Future<SyncApplyDiagnostics?> _handleEntryLink({
-    required Event event,
-    required SyncEntryLink syncMessage,
-    required JournalDb journalDb,
-  }) async {
-    final entryLink = syncMessage.entryLink;
-    final originatingHostId = syncMessage.originatingHostId;
-    final coveredVectorClocks = syncMessage.coveredVectorClocks;
-
-    final rows = await journalDb.upsertEntryLink(entryLink);
-    try {
-      if (rows > 0) {
-        _trace(
-          'apply entryLink from=${entryLink.fromId} to=${entryLink.toId} '
-          'rows=$rows',
-          subDomain: 'processor.apply.entryLink',
-        );
-      }
-    } catch (_) {
-      // best-effort logging only
-    }
-
-    // Surface DB-apply diagnostics to the pipeline when available
-    if (applyObserver != null) {
-      try {
-        final diag = SyncApplyDiagnostics(
-          eventId: event.eventId,
-          payloadType: 'entryLink',
-          vectorClock: null,
-          conflictStatus: rows == 0 ? 'entryLink.noop' : 'applied',
-          applied: rows > 0,
-          skipReason: rows > 0 ? null : JournalUpdateSkipReason.olderOrEqual,
-        );
-        applyObserver!.call(diag);
-      } catch (_) {
-        // best-effort only
-      }
-    }
-    _updateNotifications.notify(
-      {entryLink.fromId, entryLink.toId},
-      fromSync: true,
-    );
-
-    // Record in sequence log for gap detection (self-healing sync)
-    if (_sequenceLogService != null &&
-        entryLink.vectorClock != null &&
-        originatingHostId != null) {
-      final linkExists =
-          rows > 0 || await journalDb.entryLinkById(entryLink.id) != null;
-      if (linkExists) {
-        try {
-          final gaps = await _sequenceLogService.recordReceivedEntryLink(
-            linkId: entryLink.id,
-            vectorClock: entryLink.vectorClock!,
-            originatingHostId: originatingHostId,
-            coveredVectorClocks: coveredVectorClocks,
-          );
-          if (gaps.isNotEmpty) {
-            _trace(
-              'apply.entryLink.gapsDetected count=${gaps.length} '
-              'for link=${entryLink.id}',
-              subDomain: 'processor.gapDetection',
-            );
-          }
-        } catch (e, st) {
-          _loggingService.captureException(
-            e,
-            domain: 'SYNC_SEQUENCE',
-            subDomain: 'recordReceived',
-            stackTrace: st,
-          );
-        }
-      }
-    }
-    return null;
-  }
-
-  /// Resolves an agent payload from a sync message: inline first, then
-  /// fetches from [AttachmentIndex] descriptor (like [SmartJournalEntityLoader]
-  /// does for journal entities), falling back to disk.
-  ///
-  /// Agent entity files can be updated in-place (e.g. ChangeSetEntity
-  /// pending → resolved), so reading from disk alone risks stale data when
-  /// the file download hasn't completed yet. Fetching from the descriptor
-  /// ensures we always get the version that matches this text event.
-  ///
-  /// Path-validation errors from [resolveJsonCandidateFile] (e.g. path
-  /// traversal) are permanent — logged and skipped. File-read
-  /// [FileSystemException]s are rethrown so the pipeline retries (attachment
-  /// may not have arrived yet). Other exceptions (corrupt JSON, parse errors)
-  /// are logged and return null to skip permanently.
-  Future<T?> _resolveAgentPayload<T>({
-    required T? inline,
-    required String? jsonPath,
-    required T Function(Map<String, dynamic>) fromJson,
-    required String typeName,
-  }) async {
-    if (inline != null) return inline;
-    final jp = jsonPath;
-    if (jp == null) {
-      _trace(
-        '$typeName.skipped no payload and no jsonPath',
-        subDomain: 'processor.resolve',
-      );
-      return null;
-    }
-    // Validate path first — throws FileSystemException for path traversal.
-    // This is a permanent error (malformed jsonPath), so catch and skip.
-    final File file;
-    try {
-      file = resolveJsonCandidateFile(jp);
-    } on FileSystemException catch (e, st) {
-      _loggingService.captureException(
-        e,
-        domain: 'AGENT_SYNC',
-        subDomain: 'resolve.$typeName.invalidPath',
-        stackTrace: st,
-      );
-      return null;
-    }
-
-    // Fetch from the AttachmentIndex descriptor first to avoid reading
-    // stale data from disk. Agent entity files can be updated in-place
-    // (e.g. ChangeSetEntity pending → resolved), and the background
-    // download may not have completed yet when this text event arrives.
-    final fetched = await _fetchFromDescriptor(
-      jsonPath: jp,
-      targetFile: file,
-      typeName: typeName,
-    );
-    if (fetched != null) {
-      try {
-        return fromJson(json.decode(fetched) as Map<String, dynamic>);
-      } catch (e, st) {
-        _loggingService.captureException(
-          e,
-          domain: 'AGENT_SYNC',
-          subDomain: 'resolve.$typeName.parseFetched',
-          stackTrace: st,
-        );
-        return null;
-      }
-    }
-
-    // No descriptor available — fall back to disk.
-    try {
-      final jsonString = await file.readAsString();
-      return fromJson(json.decode(jsonString) as Map<String, dynamic>);
-    } on FileSystemException {
-      // Attachment file not yet available — rethrow so the pipeline retries
-      // and registers the pending descriptor path for catch-up.
-      rethrow;
-    } catch (e, st) {
-      _loggingService.captureException(
-        e,
-        domain: 'AGENT_SYNC',
-        subDomain: 'resolve.$typeName',
-        stackTrace: st,
-      );
-      return null;
-    }
-  }
-
-  /// Fetches fresh JSON from the [AttachmentIndex] descriptor and writes it
-  /// to [targetFile]. Returns the JSON string on success, or null if no
-  /// descriptor is available (index missing or not initialized).
-  ///
-  /// When a descriptor IS found but download/decode fails, throws
-  /// [FileSystemException] to prevent falling back to potentially stale
-  /// disk data.
-  Future<String?> _fetchFromDescriptor({
-    required String jsonPath,
-    required File targetFile,
-    required String typeName,
-  }) async {
-    final index = _attachmentIndex;
-    if (index == null) return null;
-
-    final indexKey = _buildAgentIndexKey(jsonPath);
-    final descriptorEvent = index.find(indexKey);
-    if (descriptorEvent == null) {
-      _trace(
-        '$typeName.descriptor.miss path=$jsonPath key=$indexKey',
-        subDomain: 'processor.resolve',
-      );
-      return null;
-    }
-
-    final dedupeKey = '$indexKey@${descriptorEvent.eventId}';
-    final existing = _inFlightDescriptorFetches[dedupeKey];
-    if (existing != null) {
-      return existing;
-    }
-
-    final future = _runDescriptorFetch(
-      jsonPath: jsonPath,
-      targetFile: targetFile,
-      typeName: typeName,
-      descriptorEvent: descriptorEvent,
-    );
-    _inFlightDescriptorFetches[dedupeKey] = future;
-    return future.whenComplete(() {
-      _inFlightDescriptorFetches.remove(dedupeKey);
-    });
-  }
-
-  Future<String?> _runDescriptorFetch({
-    required String jsonPath,
-    required File targetFile,
-    required String typeName,
-    required Event descriptorEvent,
-  }) async {
-    try {
-      final matrixFile = await downloadAttachmentWithTimeout(
-        descriptorEvent,
-        pathForError: jsonPath,
-      );
-      final downloadedBytes = matrixFile.bytes;
-      if (downloadedBytes.isEmpty) {
-        throw const FileSystemException('empty attachment bytes');
-      }
-      final bytes = await decodeAttachmentBytes(
-        event: descriptorEvent,
-        downloadedBytes: downloadedBytes,
-        relativePath: jsonPath,
-        logging: _loggingService,
-      );
-      final jsonString = utf8.decode(bytes);
-      await saveJson(targetFile.path, jsonString);
-      _trace(
-        '$typeName.descriptor.fetched path=$jsonPath bytes=${bytes.length}',
-        subDomain: 'processor.resolve',
-      );
-      return jsonString;
-    } catch (e, st) {
-      _loggingService.captureException(
-        e,
-        domain: 'AGENT_SYNC',
-        subDomain: 'resolve.$typeName.descriptorFetch',
-        stackTrace: st,
-      );
-      // Descriptor was found but download/decode failed — throw to prevent
-      // falling back to potentially stale disk data. The pipeline will retry.
-      throw FileSystemException(
-        '$typeName descriptor fetch failed',
-        jsonPath,
-      );
-    }
-  }
-
-  static String _buildAgentIndexKey(String rawPath) =>
-      normalizeAttachmentIndexKey(rawPath);
-
-  Future<AgentDomainEntity?> _resolveAgentEntity(
-    SyncAgentEntity msg,
-  ) => _resolveAgentPayload(
-    inline: msg.agentEntity,
-    jsonPath: msg.jsonPath,
-    fromJson: AgentDomainEntity.fromJson,
-    typeName: 'agentEntity',
-  );
-
-  Future<AgentLink?> _resolveAgentLink(SyncAgentLink msg) =>
-      _resolveAgentPayload(
-        inline: msg.agentLink,
-        jsonPath: msg.jsonPath,
-        fromJson: AgentLink.fromJson,
-        typeName: 'agentLink',
-      );
-
-  /// Resolves a [SyncOutboxBundle]'s manifest payload, materializes each
-  /// child's `JournalEntity` JSON to the cache on disk, and returns the
-  /// reconstructed bundle for the unpacker to walk through prepareChild.
-  ///
-  /// The manifest is the single artifact on the wire: it carries every
-  /// child's envelope plus, for `SyncJournalEntity`, the full entity body
-  /// (the database is the system of record on both ends, so the receiver
-  /// does not also need a per-child file event). Inline-payload families
-  /// (`SyncEntryLink`, `SyncAiConfig`, `SyncAiConfigDelete`,
-  /// `SyncEntityDefinition`, `SyncThemingSelection`, `SyncBackfillRequest`,
-  /// `SyncBackfillResponse`) ship their data inside the freezed envelope
-  /// and need no separate `payload` field. Agent envelopes
-  /// (`SyncAgentEntity`, `SyncAgentLink`) likewise carry their data inline.
-  ///
-  /// Vector-clock dominance is checked against the database in **one** bulk
-  /// query (no N+1) before any disk write: when the local copy already
-  /// covers the incoming envelope's clock, the on-disk JSON cache is left
-  /// untouched so the apply pipeline reads the canonical local entity
-  /// instead of an older bundled payload. The envelope is still surfaced to
-  /// the unpacker so duplicate detection and sequence-log accounting run as
-  /// they do for individually-delivered entities.
-  Future<SyncOutboxBundle?> _resolveOutboxBundleManifest(
-    String? jsonPath,
-  ) async {
-    // Phase-by-phase timing so a single grep on
-    // `processor.resolve.outboxBundle.timing` answers "where is the time
-    // going" without guesswork. Numbers in milliseconds, comparable
-    // across runs and across hosts. Captured into one log line per
-    // bundle so the typical observation is one row per drain — easy to
-    // tail or pull into a spreadsheet.
-    final totalSw = Stopwatch()..start();
-    var fetchMs = 0;
-    var diskFallbackMs = 0;
-    var decodeMs = 0;
-    var dbLookupMs = 0;
-    var writeMs = 0;
-    var manifestBytes = 0;
-    var entryCount = 0;
-    var writeCount = 0;
-    var fromDescriptor = false;
-
-    final jp = jsonPath;
-    if (jp == null) {
-      _trace(
-        'outboxBundle.skipped no jsonPath',
-        subDomain: 'processor.resolve.outboxBundle',
-      );
-      return null;
-    }
-
-    final File targetFile;
-    try {
-      targetFile = resolveJsonCandidateFile(jp);
-    } on FileSystemException catch (e, st) {
-      _loggingService.captureException(
-        e,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'processor.resolve.outboxBundle.invalidPath',
-        stackTrace: st,
-      );
-      return null;
-    }
-
-    // Reuse the existing descriptor pipeline: the encoding header on the
-    // upload tells `decodeAttachmentBytes` to gunzip the manifest bytes
-    // transparently, so what comes back is the plain manifest JSON string.
-    String? manifestJson;
-    final fetchSw = Stopwatch()..start();
-    manifestJson = await _fetchFromDescriptor(
-      jsonPath: jp,
-      targetFile: targetFile,
-      typeName: 'outboxBundle',
-    );
-    fetchMs = fetchSw.elapsedMilliseconds;
-    fromDescriptor = manifestJson != null;
-
-    if (manifestJson == null) {
-      // Fall back to disk if the descriptor is not yet registered (text
-      // event arrived before the file event made it through catch-up).
-      final diskSw = Stopwatch()..start();
-      try {
-        manifestJson = await targetFile.readAsString();
-      } on FileSystemException {
-        rethrow;
-      } catch (e, st) {
-        _loggingService.captureException(
-          e,
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'processor.resolve.outboxBundle.diskRead',
-          stackTrace: st,
-        );
-        return null;
-      }
-      diskFallbackMs = diskSw.elapsedMilliseconds;
-    }
-    manifestBytes = manifestJson.length;
-
-    final Map<String, dynamic> manifest;
-    final decodeSw = Stopwatch()..start();
-    try {
-      // Offload large manifests to a worker isolate; small ones (test
-      // fixtures, near-empty bundles) parse inline so the threshold-aware
-      // helper saves the isolate spin-up cost when the saved time is
-      // smaller than the round-trip overhead.
-      final decoded = await decodeJsonStringMaybeIsolate(manifestJson);
-      manifest = decoded! as Map<String, dynamic>;
-    } catch (e, st) {
-      _loggingService.captureException(
-        e,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'processor.resolve.outboxBundle.parse',
-        stackTrace: st,
-      );
-      return null;
-    }
-    decodeMs = decodeSw.elapsedMilliseconds;
-
-    final version = manifest['version'];
-    if (version != SyncTuning.outboxBundleManifestVersion) {
-      _loggingService.captureException(
-        'outboxBundle manifest version=$version unsupported '
-        '(expected ${SyncTuning.outboxBundleManifestVersion}) — skipping',
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'processor.resolve.outboxBundle.unknownVersion',
-      );
-      return null;
-    }
-
-    final rawEntries = manifest['entries'];
-    if (rawEntries is! List) {
-      _loggingService.captureException(
-        'outboxBundle manifest missing entries array',
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'processor.resolve.outboxBundle.malformed',
-      );
-      return null;
-    }
-
-    // First pass: parse envelopes and collect every SyncJournalEntity id we
-    // will need a local copy of, so the bulk DB fetch below stays a single
-    // round-trip even at the [SyncTuning.outboxBundleMaxSize] cap.
-    final parsedEntries = <_OutboxBundleManifestEntry>[];
-    final journalEntityIds = <String>{};
-    for (final raw in rawEntries) {
-      if (raw is! Map<String, dynamic>) continue;
-      final envelopeJson = raw['envelope'];
-      if (envelopeJson is! Map<String, dynamic>) continue;
-      final SyncMessage envelope;
-      try {
-        envelope = SyncMessage.fromJson(envelopeJson);
-      } catch (e, st) {
-        _loggingService.captureException(
-          e,
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'processor.resolve.outboxBundle.envelopeParse',
-          stackTrace: st,
-        );
-        continue;
-      }
-      if (envelope is SyncOutboxBundle) {
-        // Defence in depth: nested bundles are rejected by the unpacker
-        // too, but skipping here saves work and avoids a bogus DB lookup.
-        _trace(
-          'outboxBundle.entry.skipNested',
-          subDomain: 'processor.resolve.outboxBundle',
-        );
-        continue;
-      }
-      parsedEntries.add(
-        _OutboxBundleManifestEntry(envelope: envelope, payload: raw['payload']),
-      );
-      if (envelope is SyncJournalEntity) {
-        journalEntityIds.add(envelope.id);
-      }
-    }
-
-    entryCount = parsedEntries.length;
-
-    // VC-dominance check uses the database as system of record. If no
-    // JournalDb was injected (legacy test harness), the dominance check is
-    // skipped and the incoming JSON is always written — safe but loses the
-    // "don't clobber a fresher local cache" optimization.
-    final db = _journalDb;
-    final dbSw = Stopwatch()..start();
-    final localEntities = journalEntityIds.isEmpty || db == null
-        ? const <String, JournalEntity>{}
-        : await db.journalEntityMapForIds(journalEntityIds);
-    dbLookupMs = dbSw.elapsedMilliseconds;
-
-    // Plan every disk write before issuing any of them so we can fail the
-    // whole bundle on a malformed manifest (e.g. a SyncJournalEntity with
-    // no `payload`) without leaving half the entries on disk. Per-write
-    // FileSystemExceptions are surfaced via the rethrow below so the
-    // unpacker's IOException-rethrow path triggers a whole-bundle retry —
-    // already-applied writes stay (idempotent under VC dominance on the
-    // next pass).
-    final writePlans = <_OutboxBundleWritePlan>[];
-    final children = <SyncMessage>[];
-    for (final parsed in parsedEntries) {
-      final envelope = parsed.envelope;
-      final payload = parsed.payload;
-
-      if (envelope is SyncJournalEntity) {
-        final local = localEntities[envelope.id];
-        final localVc = local?.meta.vectorClock;
-        final incomingVc = envelope.vectorClock;
-        var localDominates = false;
-        if (local != null && localVc != null && incomingVc != null) {
-          final cmp = VectorClock.compare(localVc, incomingVc);
-          localDominates =
-              cmp == VclockStatus.a_gt_b || cmp == VclockStatus.equal;
-        }
-
-        final Map<String, dynamic> payloadToWrite;
-        if (localDominates) {
-          // Refresh the on-disk cache from the canonical DB version. The
-          // cache file may be stale or missing — writing the dominant
-          // local entity back guarantees the apply pipeline's
-          // SmartJournalEntityLoader reads the right data instead of
-          // serving a stale fixture or hitting a descriptor miss.
-          // `local` is non-null here because `localDominates` requires it.
-          payloadToWrite = local!.toJson();
-          _trace(
-            'outboxBundle.entry.localDominates id=${envelope.id} '
-            'jsonPath=${envelope.jsonPath}',
-            subDomain: 'processor.resolve.outboxBundle',
-          );
-        } else {
-          if (payload is! Map<String, dynamic>) {
-            // Malformed manifest: a SyncJournalEntity envelope without an
-            // inline payload cannot be applied. Skipping this child would
-            // ack the bundle while the entity never reaches the local DB
-            // (silent loss). Drop the whole bundle instead — peers can
-            // recover the entry via the sequence-log backfill path.
-            _loggingService.captureException(
-              'outboxBundle entry missing payload for SyncJournalEntity '
-              'id=${envelope.id} jsonPath=${envelope.jsonPath} — '
-              'dropping the whole bundle so missing entries surface via '
-              'the sequence-log backfill path instead of being lost',
-              domain: 'MATRIX_SERVICE',
-              subDomain: 'processor.resolve.outboxBundle.missingPayload',
-            );
-            return null;
-          }
-          payloadToWrite = payload;
-        }
-
-        final File entityFile;
-        try {
-          entityFile = resolveJsonCandidateFile(envelope.jsonPath);
-        } on FileSystemException catch (e, st) {
-          _loggingService.captureException(
-            e,
-            domain: 'MATRIX_SERVICE',
-            subDomain: 'processor.resolve.outboxBundle.invalidEntryPath',
-            stackTrace: st,
-          );
-          return null;
-        }
-
-        writePlans.add(
-          _OutboxBundleWritePlan(
-            envelopeId: envelope.id,
-            jsonPath: envelope.jsonPath,
-            filePath: entityFile.path,
-            payload: payloadToWrite,
-          ),
-        );
-      }
-
-      children.add(envelope);
-    }
-
-    // Issue per-child writes serially. The earlier parallel `Future.wait`
-    // multiplied the prepare phase's per-batch fan-out by up to
-    // [outboxBundleMaxSize], which on phone filesystems made the inbound
-    // worker stall under heavy backfill — too many concurrent atomic
-    // writes (each one is parent-create + temp + rename) ran into FD caps
-    // and journal backpressure, and the resulting FileSystemException
-    // pushed the bundle straight back into a tight retry loop with no
-    // forward progress. Sequential writes within a bundle keep the worst-
-    // case write fan-out bounded by the queue's batch concurrency
-    // (`SyncTuning.queueBatchSize`) rather than batch × per-bundle.
-    final writeSw = Stopwatch()..start();
-    for (final plan in writePlans) {
-      await saveJson(plan.filePath, json.encode(plan.payload));
-      _trace(
-        'outboxBundle.entry.materialized id=${plan.envelopeId} '
-        'jsonPath=${plan.jsonPath}',
-        subDomain: 'processor.resolve.outboxBundle',
-      );
-    }
-    writeMs = writeSw.elapsedMilliseconds;
-    writeCount = writePlans.length;
-
-    _trace(
-      'totalMs=${totalSw.elapsedMilliseconds} '
-      'fetchMs=$fetchMs '
-      'fromDescriptor=$fromDescriptor '
-      'diskFallbackMs=$diskFallbackMs '
-      'manifestBytes=$manifestBytes '
-      'decodeMs=$decodeMs '
-      'entries=$entryCount '
-      'dbLookupMs=$dbLookupMs '
-      'dbHits=${localEntities.length}/${journalEntityIds.length} '
-      'writeMs=$writeMs '
-      'writes=$writeCount',
-      subDomain: 'processor.resolve.outboxBundle.timing',
-    );
-
-    return SyncOutboxBundle(
-      children: children,
-      jsonPath: jp,
-    );
-  }
-
   @visibleForTesting
   Future<SyncOutboxBundle?> resolveOutboxBundleManifestForTesting(
     String? jsonPath,
@@ -1474,175 +454,6 @@ class SyncEventProcessor {
     loggingService: _loggingService,
     trace: _trace,
   );
-
-  Future<void> _applyAgentEntityMessage({
-    required SyncAgentEntity msg,
-    required AgentDomainEntity? resolvedEntity,
-  }) async {
-    if (resolvedEntity == null) {
-      return;
-    }
-    if (agentRepository != null) {
-      await agentRepository!.upsertEntity(resolvedEntity);
-      // Remove wake subscriptions when an agent is paused or destroyed
-      // remotely — mirrors what AgentService.pauseAgent/destroyAgent do
-      // locally.
-      if (wakeOrchestrator != null &&
-          resolvedEntity is AgentIdentityEntity &&
-          resolvedEntity.lifecycle != AgentLifecycle.active) {
-        wakeOrchestrator!.removeSubscriptions(resolvedEntity.agentId);
-      }
-      // Restore wake subscriptions when an agent is resumed remotely —
-      // mirrors what TaskAgentService.restoreSubscriptionsForAgent does
-      // locally after AgentService.resumeAgent.
-      if (wakeOrchestrator != null &&
-          resolvedEntity is AgentIdentityEntity &&
-          resolvedEntity.lifecycle == AgentLifecycle.active &&
-          resolvedEntity.kind == 'task_agent') {
-        final links = await agentRepository!.getLinksFrom(
-          resolvedEntity.agentId,
-          type: 'agent_task',
-        );
-        for (final link in links) {
-          wakeOrchestrator!.addSubscription(
-            AgentSubscription(
-              id: '${resolvedEntity.agentId}_task_${link.toId}',
-              agentId: resolvedEntity.agentId,
-              matchEntityIds: {link.toId},
-            ),
-          );
-        }
-      }
-      _updateNotifications.notify(
-        {
-          resolvedEntity.agentId,
-          // Include templateId so template-level aggregate providers
-          // refresh when token usage or reports arrive from other devices.
-          if (resolvedEntity is WakeTokenUsageEntity &&
-              resolvedEntity.templateId != null)
-            resolvedEntity.templateId!,
-          agentNotification,
-        },
-        fromSync: true,
-      );
-      _trace(
-        'apply agentEntity id=${resolvedEntity.id}',
-        subDomain: 'processor.apply',
-      );
-
-      // Record in sequence log for gap detection (self-healing sync)
-      if (_sequenceLogService != null &&
-          resolvedEntity.vectorClock != null &&
-          msg.originatingHostId != null) {
-        try {
-          final gaps = await _sequenceLogService.recordReceivedEntry(
-            entryId: resolvedEntity.id,
-            vectorClock: resolvedEntity.vectorClock!,
-            originatingHostId: msg.originatingHostId!,
-            coveredVectorClocks: msg.coveredVectorClocks,
-            payloadType: SyncSequencePayloadType.agentEntity,
-            jsonPath: msg.jsonPath,
-          );
-          if (gaps.isNotEmpty) {
-            _trace(
-              'apply.agentEntity.gapsDetected count=${gaps.length} '
-              'for entity=${resolvedEntity.id}',
-              subDomain: 'processor.gapDetection',
-            );
-          }
-        } catch (e, st) {
-          _loggingService.captureException(
-            e,
-            domain: 'SYNC_SEQUENCE',
-            subDomain: 'recordReceived',
-            stackTrace: st,
-          );
-        }
-      }
-    } else {
-      _trace(
-        'agentEntity.ignored no repository',
-        subDomain: 'processor.apply',
-      );
-    }
-  }
-
-  Future<void> _applyAgentLinkMessage({
-    required SyncAgentLink msg,
-    required AgentLink? resolvedLink,
-  }) async {
-    if (resolvedLink == null) {
-      return;
-    }
-    if (agentRepository != null) {
-      await agentRepository!.upsertLink(resolvedLink);
-      // Restore wake subscription when an agent_task link arrives for an
-      // active task_agent. This handles the case where the link arrives
-      // after the identity — the SyncAgentEntity handler queries existing
-      // links, which may be empty if the link hasn't been synced yet.
-      // addSubscription is idempotent (replaces by ID), so both handlers
-      // firing for the same agent is harmless.
-      if (wakeOrchestrator != null &&
-          resolvedLink is AgentTaskLink &&
-          resolvedLink.deletedAt == null) {
-        final agent = await agentRepository!.getEntity(resolvedLink.fromId);
-        if (agent is AgentIdentityEntity &&
-            agent.lifecycle == AgentLifecycle.active &&
-            agent.kind == 'task_agent') {
-          wakeOrchestrator!.addSubscription(
-            AgentSubscription(
-              id: '${resolvedLink.fromId}_task_${resolvedLink.toId}',
-              agentId: resolvedLink.fromId,
-              matchEntityIds: {resolvedLink.toId},
-            ),
-          );
-        }
-      }
-      _updateNotifications.notify(
-        {resolvedLink.fromId, resolvedLink.toId, agentNotification},
-        fromSync: true,
-      );
-      _trace(
-        'apply agentLink id=${resolvedLink.id}',
-        subDomain: 'processor.apply',
-      );
-
-      // Record in sequence log for gap detection (self-healing sync)
-      if (_sequenceLogService != null &&
-          resolvedLink.vectorClock != null &&
-          msg.originatingHostId != null) {
-        try {
-          final gaps = await _sequenceLogService.recordReceivedEntry(
-            entryId: resolvedLink.id,
-            vectorClock: resolvedLink.vectorClock!,
-            originatingHostId: msg.originatingHostId!,
-            coveredVectorClocks: msg.coveredVectorClocks,
-            payloadType: SyncSequencePayloadType.agentLink,
-            jsonPath: msg.jsonPath,
-          );
-          if (gaps.isNotEmpty) {
-            _trace(
-              'apply.agentLink.gapsDetected count=${gaps.length} '
-              'for link=${resolvedLink.id}',
-              subDomain: 'processor.gapDetection',
-            );
-          }
-        } catch (e, st) {
-          _loggingService.captureException(
-            e,
-            domain: 'SYNC_SEQUENCE',
-            subDomain: 'recordReceived',
-            stackTrace: st,
-          );
-        }
-      }
-    } else {
-      _trace(
-        'agentLink.ignored no repository',
-        subDomain: 'processor.apply',
-      );
-    }
-  }
 
   Future<SyncApplyDiagnostics?> _applyMessage({
     required PreparedSyncEvent prepared,
@@ -1936,33 +747,4 @@ class PreparedSyncEvent {
   /// apply can recurse through the existing per-type pipeline. Null means
   /// the bundle payload could not be resolved and apply will skip it.
   final PreparedOutboxSyncBundle? resolvedOutboxBundle;
-}
-
-/// One parsed entry from a [SyncOutboxBundle]'s manifest: the deserialized
-/// envelope plus its inline payload (only present for [SyncJournalEntity]
-/// children — agent and inline-only families carry their data inside the
-/// envelope itself).
-class _OutboxBundleManifestEntry {
-  _OutboxBundleManifestEntry({required this.envelope, required this.payload});
-
-  final SyncMessage envelope;
-  final Object? payload;
-}
-
-/// One pending on-disk JSON cache write planned by the outbox bundle
-/// resolver. The resolver builds a list of plans up-front (rejecting the
-/// whole bundle on malformed entries) and dispatches them via
-/// `Future.wait` so 50 atomic writes overlap instead of running serially.
-class _OutboxBundleWritePlan {
-  _OutboxBundleWritePlan({
-    required this.envelopeId,
-    required this.jsonPath,
-    required this.filePath,
-    required this.payload,
-  });
-
-  final String envelopeId;
-  final String jsonPath;
-  final String filePath;
-  final Map<String, dynamic> payload;
 }
